Refactoring AssetStatus to tsx
